convert signup to middleware function; chain customAuth after signup
